requirements.txt
tweepy
nltk
pandas
requests
beautifulsoup4
discord.py
requests >= 2.26.0
pytest >= 6.2.4
pytest-asyncio >= 0.15.1
aiohttp >= 3.8.4
cryptography >= 41.0.2
instagrapi >= 1.17.13
pillow >= 10.0.0
colorama >= 0.4.6